Refactor server code into more packages
Create new packages:

* pkg/common (for common types)
* pkg/integration (for code that will integrate metalkube components
  into facet)

The routes.go file now only contains router code.  It's the primary
place for connecting routes to handlers.

All handlers have been moved to handlers.go.

The HostsHandler has been refactor to call a function which is
responsible for actually retrieving the hosts information.  This
function is called GetHosts and it lives in the integration package.