MGMT-8715: uses UBI images and embeds nginx config
Embedding the nginx config file eliminates the error-prone process of
making the user mount in a customized config file at runtime. The image
now takes an environment variable for the one setting that matters.

Use of UBI images brings this project closer to the rest of the
openshift ecosystem, where the entire stack is based on el8. In
particular, offering a container-based build process eliminates any
variability that can be introduced by changes in an unknown and/or
untracked external build system.